Update meta.yaml
Updating meta.yml for conda build.  Related to: https://github.com/opensim-org/opensim-core/issues/3273

Recommended [here](https://docs.conda.io/projects/conda-build/en/latest/resources/define-metadata.html#run) that the numpy version should be specified under the requirements/run section. This is where we specify the requirements for the runtime environment. They also specified something about specifying `--numpy` with `conda-build` if using this syntax - I haven't built recipes before so I'll leave that part for someone else that knows the usual syntax/where/when to call this. 

I also deleted the build string because I believe that this should be done automatically if you specify a version for requirements/run (but Im not positive on this one). I think that this string should be built by conda-build to have a unique hash that goes after the python version to specify the dependencies etc. and this is how different requirements (e.g., numpy) are specified.